From Stavros:
it's a bit inconvenient to have the models in one directory and the
properties in another. also properties are called \"experiments\",
which seems strange. i would put the entire TCP example in one directory,
and rename properties property1.pml, property2.pml, etc.

I made the requested change but did not yet test. - Max